+++ paper_uploads/signals/__init__.py
@@ -2,7 +2,7 @@ from django.db.models.signals import Signal
 
 # Хэш файла изменился. Это происходит либо в случае создания нового файла,
 # либо при изменении контента уже существующего файла.
-hash_updated = Signal(providing_args=["instance", "hash"])
+content_hash_update = Signal(providing_args=["instance", "content_hash"])
 
 # создание / перезапись (recut) файла вариации изображения
 variation_created = Signal(providing_args=["instance", "file"])