+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [0.3.0](https://github.com/dldevinc/paper-uploads/tree/v0.3.0) - 2020-07-22
+- Rename `hash_updated` signal to `content_hash_update`
 - Signal `variation_created` was added
 - Remove postprocessing
 - `utils.run_validators()` and `utils.get_instance()` was moved to `helpers.py`