+++ CHANGELOG.md
@@ -1,11 +1,10 @@
 # Change Log
 
 ## [0.3.0](https://github.com/dldevinc/paper-uploads/tree/v0.3.0) - 2020-08-04
+- Rename `ItemField` to `CollectionItem`
 - Rename `MimetypeValidator` to `MimeTypeValidator`
 - Signal `variation_created` was added
-- `utils.run_validators()` and `utils.get_instance()` was moved to `helpers.py`
-- `utils.lowercase_copy()` renamed to `utils.lowercased_dict_keys()`
-- Use DropBox checksum realization
+- Migrate to DropBox's checksum realization
 - Remove postprocessing
 - Refactored class hierarchy
 - Upgrade development environment