+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Change Log
 
-## [0.3.0](https://github.com/dldevinc/paper-uploads/tree/v0.3.0) - 2020-08-04
+## [0.4.0](https://github.com/dldevinc/paper-uploads/tree/v0.4.0) - 2020-09-04
+- **Rewritten from scratch**
+- Added `CLOUDINARY_TYPE` setting
+- Added `CLOUDINARY_TEMP_DIR` setting
+- Rename `CLOUDINARY` setting to `CLOUDINARY_UPLOADER_OPTIONS`
 - Rename `ItemField` to `CollectionItem`
 - Rename `MimetypeValidator` to `MimeTypeValidator`
 - Signal `variation_created` was added
 - Migrate to DropBox's checksum realization
 - Remove postprocessing
-- Refactored class hierarchy
 - Upgrade development environment